requirements-dev.txt
# Hermes Development Dependencies

# Include production dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-docker>=2.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
isort>=5.13.0

# Pre-commit hooks
pre-commit>=3.5.0

# Type stubs
types-PyYAML>=6.0.0
types-python-dateutil>=2.8.0
